test(web): extract integration tests behind build tag
Move TestAnalyzeSSE_Diagnosis and TestAnalyzeSSE_PassingRun to
analyze_integration_test.go with //go:build integration. Unit tests
run by default, integration tests require -tags=integration.
Revert -short in lefthook pre-push (no longer needed).
