+++ .lefthook.yml
@@ -12,4 +12,4 @@ pre-push:
   parallel: true
   commands:
     test:
-      run: go test -short -race ./...
+      run: go test -race ./...
